Move parse code to open rosa module
