+++ collect_app/src/main/java/org/odk/collect/android/formmanagement/FormSourceProvider.kt
@@ -1,10 +1,10 @@
 package org.odk.collect.android.formmanagement
 
 import org.odk.collect.android.openrosa.OpenRosaFormSource
-import org.odk.collect.android.openrosa.OpenRosaResponseParserImpl
 import org.odk.collect.android.utilities.WebCredentialsUtils
 import org.odk.collect.forms.FormSource
 import org.odk.collect.openrosa.http.OpenRosaHttpInterface
+import org.odk.collect.openrosa.parse.OpenRosaResponseParserImpl
 import org.odk.collect.projects.ProjectDependencyFactory
 import org.odk.collect.settings.keys.ProjectKeys
 import org.odk.collect.shared.settings.Settings
